Cap minimum fetched messages at 20
We were seeing endless requests being made for very small numbers of
messages to the endpoint when purging. This should mitigate that to
some extent.

+++ fallacy.go
@@ -128,7 +128,7 @@ func (f *Fallacy) HandleMessage(ev *gomatrix.Event) {
 			f.printHelp(ev.RoomID)
 			return
 		}
-		if err := f.PurgeMessages(ev.RoomID, "", int32(l)); err != nil {
+		if err := f.PurgeMessages(ev.RoomID, "", l); err != nil {
 			log.Println(err)
 		}
 		return
